# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MemeHub API - 9GAG Clone
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get posts feed
      parameters:
        - name: section
          in: query
          schema:
            type: string
            enum: [hot, trending, fresh, top]
            default: hot
        - name: tag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 10
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsListResponse"

    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /posts/{id}/vote:
    post:
      operationId: votePost
      tags: [votes]
      summary: Upvote or downvote a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteRequest"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResponse"
  /posts/{id}/save:
    post:
      operationId: savePost
      tags: [posts]
      summary: Save or unsave a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Save toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveResponse"
...
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
            default: 1
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsListResponse"
...
  /admin/posts:
    get:
      operationId: getAdminPosts
      tags: [admin]
      summary: Get all posts (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, removed]
            default: all
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPostsListResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Post:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        gifUrl:
          type: string
          nullable: true
        type:
          type: string
          enum: [image, gif, video]
        upvotes:
          type: number
        downvotes:
          type: number
        points:
          type: number
        commentsCount:
          type: number
        viewsCount:
          type: number
        status:
          type: string
          enum: [pending, approved, removed]
        userVote:
          type: string
          enum: [up, down]
          nullable: true
        isSaved:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        author:
          $ref: "#/components/schemas/PostAuthor"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - imageUrl
        - type
        - upvotes
        - downvotes
        - points
        - commentsCount
        - viewsCount
        - status
        - isSaved
        - tags
        - author
        - createdAt
    PostAuthor:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatar:
          type: string
          nullable: true
      required:
        - id
        - username
    PostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
        hasMore:
          type: boolean
      required:
        - posts
        - total
        - page
        - totalPages
        - hasMore
...
    VoteRequest:
      type: object
      properties:
        direction:
          type: string
          enum: [up, down]
      required:
        - direction
    VoteResponse:
      type: object
      properties:
        upvotes:
          type: number
        downvotes:
          type: number
        points:
          type: number
        userVote:
          type: string
          enum: [up, down]
          nullable: true
      required:
        - upvotes
        - downvotes
        - points
    SaveResponse:
      type: object
      properties:
        isSaved:
          type: boolean
      required:
        - isSaved
...
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        postsCount:
          type: number
        color:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
        - postsCount
...
    CreatePostRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        imageUrl:
          type: string
        type:
          type: string
          enum: [image, gif, video]
        tagIds:
          type: array
          items:
            type: string
      required:
        - title
        - imageUrl
        - type
...
    AdminPostsListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: number
        page:
          type: number
        totalPages:
          type: number
      required:
        - posts
        - total
        - page
        - totalPages
...